Disabled Radeonboost.kext as a default
Today I helped out one of my patreons and as soon as we enabled the Radeonboost.kext (even the newest v1.6), his system didn't boot anymore. He is using 10.15.6 and has a AMD Radeon 5700 XT. And because it is causing issues for many people with the NAVI-based GPUs, I will disable this kext as a default.